To announce

+++ src/Util/Dice.java
@@ -7,8 +7,7 @@ public class Dice {
         dice_num = (int)(Math.random() * 6) + 1;
     }
 
-    public int roll_dice(){
-        //dice_num = 2;
+    public float roll_dice(){
         dice_num = (int)(Math.random() * 6) + 1;
         return dice_num;
     }
